Make remaining tests policy-aware under numpy+floatX
- scan test_no_step/test_no_steps_sit_sot: pin the inner 2*x literal to
  int8 so the state dtype is preserved and scan accepts the update
- scalar test_loop: pin inner-graph literals so the float16 rebuild
  keeps update dtype == init dtype
- scalar test_constant, numba test_rv_float_params_cast_respects_warn_float64:
  expect int64 literals under numpy+floatX
- test_downcast_dtype: pin y = x * np.float32(2) so the grad downcast
  mechanism stays exercised now that x * 2 legitimately promotes

+++ tests/scan/test_basic.py
@@ -404,7 +404,7 @@ class TestScan:
         """We expect an empty output array when ``n_steps == 0``."""
 
         def f_pow(x_tm1):
-            return 2 * x_tm1
+            return np.int8(2) * x_tm1
 
         n_steps = iscalar("n_steps")
         values = scan(
@@ -439,7 +439,7 @@ class TestScan:
         """We expect an empty output array when scanning over an empty sequence."""
 
         def inner_fn(x_seq, x_i):
-            return 2 * x_i
+            return np.int8(2) * x_i
 
         with config.change_flags(mode=mode):
             values = scan(
